fix(oidc): move preauthorized picker to login
